debugging
In a, I deleted the line which showed b=0, because denominator can't be 0.
In b, I indented the space in the last line.
In c, it can work successfully, so  think it's a correct code.
In d, I added the brackets at the end of the line.
In e, I corrected the spelling about "string" to "strng".